% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add}
\alias{add}
\title{FUNCTION_TITLE}
\usage{
add(.object, .fun, ...)
}
\arguments{
\item{.object}{PARAM_DESCRIPTION}

\item{.fun}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
