% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_heatcircle.R
\name{add_heatcircle}
\alias{add_heatcircle}
\alias{lotri_heatcircle}
\alias{utri_heatcircle}
\title{Add circles whose size represents correlation strength}
\usage{
lotri_heatcircle(
  corr_method = "pearson",
  rmin = 0.1,
  rmax = 0.9,
  scale_by = c("area", "radius"),
  ...
)

utri_heatcircle(
  corr_method = "pearson",
  rmin = 0.1,
  rmax = 0.9,
  scale_by = c("area", "radius"),
  ...
)
}
\arguments{
\item{corr_method}{character string with the correlation method passed
to \code{\link[stats]{cor}}. Can be one of "pearson", "kendall" and
"spearman". Defaults to "pearson" (or is inherited from the setting
in \code{\link[ggcorrm]{ggcorrm()}}).}

\item{rmin}{numeric (0 - 1). Minimum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.1.}

\item{rmax}{numeric (0 - 1). Maximum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.9.}

\item{scale_by}{character.  \code{scale_by = "area"} scales the area of the
circles by correlation strength, \code{scale_by = "radius"} scales by
radius. Defaults to "area".}

\item{...}{Additional arguments to \code{\link{stat_heatcircle}}.}
}
\value{
A \code{ggplot2} layer with circles displaying correlation strength.
}
\description{
\code{lotri_heatcircle()} and \code{utri_heatcircle()} are used
  to display circles whose fill and  size aesthetics depend on correlation
  strength in the lower/upper off-diagonal facets of \code{\link{ggcorrm}}
  plots.
}
\details{
\code{lotri_heatcircle()} and \code{utri_heatcircle()} are wrappers
  around \code{\link[stat_heatcircle]{stat_heatcircle()}} that additionally
  take care of the right specification of aesthetics.

 \code{stat_heatcircle()} computes the outlines of circles that
 are centered around the midpoint of \code{gcorrm} facets and whose radius
 or area is modified in dependence of correlation strength. \code{rmin} and
 \code{rmax} allow to modify the range in which the radii of the circles may
 vary.
}
\seealso{
\code{\link{stat_heatcircle}},
  \code{\link{stat_heatmap}}
}
