% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{corrmoran_themes}
\alias{corrmoran_themes}
\alias{theme_corrm}
\title{Themes for ggcorrm plots}
\usage{
theme_corrm(
  base_size = 9,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size (defaults to 9).}

\item{base_family}{base font familiy (defaults to "").}

\item{base_line_size}{base size for line elements (defaults to base_size/22).}

\item{base_rect_size}{base sie for rect elements  (defaults to base_size/22).}
}
\description{
ggplot themes controlling all non-data display meant
    for the use with \code{ggcorrm}. Single elements of a theme can
    be modified with \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\details{
Available themes:
    \describe{
        \item{theme_corrm()}{The standard corrmorant theme with white background and
              strong black facet outlines. \emph{So far the only available
              corrmorant theme.}
              }
           }
}
\seealso{
\code{\link[ggplot2:ggtheme]{ggplot2::ggtheme}},
  \code{\link[ggplot2:theme]{ggplot2::theme}},
  \code{\link{ggcorrm}},
  \code{\link{tidy_corrm}},
  \code{\link{corrmorant}}
}
