% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_ggproto.R, R/stat_corrtext.R,
%   R/stat_dia_bin.R, R/stat_dia_density.R, R/stat_dia_names.R,
%   R/stat_heatcircle.R, R/stat_heatmap.R
\docType{data}
\name{corrmorant_ggproto}
\alias{corrmorant_ggproto}
\alias{StatCorrtext}
\alias{StatDiaBin}
\alias{StatDiaDensity}
\alias{StatDiaNames}
\alias{StatHeatcircle}
\alias{StatHeatmap}
\title{ggproto classes for corrmorant}
\description{
Class definitions for the new \code{ggproto} classes defined in the
\code{corrmorant} package.
}
\section{New stats}{

\itemize{
\item \code{StatCorrtext}: create and position text labels for correlation strength
(see \link{stat_corrtext} for details).
\item \code{StatHeatmap}: create correlation heat maps (see \link{stat_heatmap} for
details).
\item \code{StatHeatcircle}: create circles indicating correlation strength (see
\link{stat_heatcircle} for details).
\item \code{StatDiaNames}: create and position text labels for variable names in the
diagonal facets of \code{ggcorrm} plots (see \link{stat_dia_names} for details).
\item \code{StatDiaDensity}: create and position density plots of variables in the
diagonal facets of \code{ggcorrm} plots (see \link{stat_dia_density} for details).
\item \code{StatDiaBin}: create and position plots of binned data (histograms or
frequency polygons) of variables in the diagonal facets of \code{ggcorrm} plots
(see \link{stat_dia_bin} for details).
}
}

\seealso{
\link{ggplot2-ggproto},
\link{lotri_corrtext} and \link{utri_corrtext},
\link{lotri_heatmap} and \link{utri_heatmap},
\link{lotri_heatcircle} and \link{utri_heatcircle},
\link{dia_names},
\link{dia_density},
\link{dia_histogram}
}
\keyword{datasets}
