% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_package.R
\docType{package}
\name{corrmorant_package}
\alias{corrmorant_package}
\title{corrmorant - Flexible correlation matrices based on ggplot2}
\description{
corrmorant extends ggplot2 by an automated framework for plots of
correlation matrices that can be easily modified via regular ggplot2 syntax.
In addition, it provides a large set of visualization tools for exploratory
data analysis based on correlation matrices.
}
\section{Author}{


\strong{Roman M. Link} (Maintainer, \link{roman.link@plant-ecology.de})
}

\section{Where to start}{

\itemize{
\item \code{\link[=corrmorant]{corrmorant()}} - a simple, stripped down function for fast vizualization
of correlation matrices.
\item \code{\link[=ggcorrm]{ggcorrm()}} - the workhorse function for all corrmorant plots.
\item \code{\link[=tidy_corrm]{tidy_corrm()}} - prepare datasets for plotting in corrmorant.
\item \link{corrmorant_selectors} - functions that modify regular \code{ggplot2} layers to
direct them to the right positions in \code{ggcorrm()} plots: \code{dia()}, \code{lotri()}
and \code{utri()}.
}
}

\section{Special corrmorant layers}{


\strong{Modifying the plot diagonal}
\itemize{
\item \code{\link[=dia_names]{dia_names()}} - add text labels for variable names.
\item \code{\link[=dia_density]{dia_density()}} - add density plots.
\item \code{\link[=dia_histogram]{dia_histogram()}} - add histograms.
\item \code{\link[=dia_freqpoly]{dia_freqpoly()}} - add frequency polygons.
\item \code{\link[=dia]{dia()}} - add arbitrary ggplot layers to the diagonal.
}

\strong{Modifying the lower triangle}
\itemize{
\item \code{\link[=lotri_corrtext]{lotri_corrtext()}} - add text labels for correlation strength.
\item \code{\link[=lotri_funtext]{lotri_funtext()}} - add text labels for user-specified functions.
\item \code{\link[=lotri_heatmap]{lotri_heatmap()}} - add correlation heatmaps.
\item \code{\link[=lotri_heatpoints]{lotri_heatpoints()}} - add symbols whose size and color indicates correlation strength.
\item \code{\link[=lotri_heatcircle]{lotri_heatcircle()}} - add circles whose area scales with correlation strength.
\item \code{\link[=lotri]{lotri()}} - add arbitrary ggplot layers to the lower triangle.
}

\strong{Modifying the upper triangle}
\itemize{
\item \code{\link[=utri_corrtext]{utri_corrtext()}} - add text labels for correlation strength.
\item \code{\link[=utri_funtext]{utri_funtext()}} - add text labels for user-specified functions.
\item \code{\link[=utri_heatmap]{utri_heatmap()}} - add correlation heatmaps.
\item \code{\link[=utri_heatpoints]{utri_heatpoints()}} - add symbols whose size and color indicates correlation strength.
\item \code{\link[=utri_heatcircle]{utri_heatcircle()}} - add circles whose area scales with correlation strength.
\item \code{\link[=utri]{utri()}} - add arbitrary ggplot layers to the upper triangle.
}
}

