% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{corrmorant_scales}
\alias{corrmorant_scales}
\alias{scale_color_corr}
\alias{scale_fill_corr}
\title{Color and fill scales for correlation plots}
\usage{
scale_color_corr(
  ...,
  option = c("A", "B", "C", "D"),
  limits = c(-1, 1),
  na.value = "#55FF55",
  guide = "colourbar",
  aesthetics = "colour",
  direction = 1
)

scale_fill_corr(
  ...,
  option = c("A", "B", "C", "D"),
  limits = c(-1, 1),
  na.value = "#55FF55",
  guide = "colourbar",
  aesthetics = "fill",
  direction = 1
)
}
\arguments{
\item{...}{arguments passed to continuous scale}

\item{option}{character in c("A", "B", "C", "D") specifying the desired
color scale. Defaults to "A".}

\item{limits}{Limits of the color gradient By default, the limits are set
to c(-1, 1), and it is likely not a wise idea to change it as it might
over-emphasise weak correlations.}

\item{na.value}{the color to use for missing values. Defaults to '#55FF55'.}

\item{guide}{Type of legend ("colourbar" for continuous legend, "legend"
for legend with discrete values). Defaults to "colourbar".}

\item{aesthetics}{Character string or vector of character strings with the
aesthetics linked to the scale.  Defaults to  "color" for
\code{scale_color_corr()} and "fill" for \code{scale_fill_corr()}.}

\item{direction}{Numeric indicating the order of the scale. Defaults to 1;
negative values reverse the order of the scale.}
}
\description{
Diverging color gradients for the illustration of correlation
strength.
}
\details{
The color and fill scales specified via \code{option} are built upon
\code{\link[=ggplot2:scale_colour_gradient2]{ggplot2:scale_colour_gradient2()}},
which can easily be used to set up additional scales if the available
options are not sufficient.
}
\examples{
\dontrun{
if(interactive()){
# iris data with conditional coloring by correlation strength
ggcorrm(iris, rescale = "by_sd") +
  lotri(geom_point(aes(col = .corr), alpha = 0.6)) +
  lotri(geom_smooth(aes(fill = .corr, col = .corr),
                    method = "lm", size = 0.3, alpha = 0.6)) +
  utri_corrtext(aes(color = .corr)) +
  dia_density(fill = "grey80", lower = .4) +
  dia_names(y_pos = .1) +
  scale_color_corr(option = "A", aesthetics = c("fill", "color"))
  }
}
}
\seealso{
\link[ggplot2:scale_color_gradient2]{ggplot2::scale_color_gradient2}
}
