% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_stats.R, R/stat_cortext.R,
%   R/stat_dia_binned.R, R/stat_dia_density.R, R/stat_dia_names.R
\name{corrmorant_stats}
\alias{corrmorant_stats}
\alias{stat_cortext}
\alias{stat_dia_bin}
\alias{stat_dia_density}
\alias{stat_dia_names}
\title{New stats for corrmorant}
\usage{
stat_cortext(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  nrow = nrow,
  ncol = ncol,
  digits = 2,
  corr_method = "pearson",
  squeeze = 0.7,
  ...
)

stat_dia_bin(
  mapping = NULL,
  data = NULL,
  geom = "rect",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  lower = 0.25,
  upper = 1,
  barwidth = 0.9,
  ...
)

stat_dia_density(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  lower = 0.25,
  upper = 1,
  ...
)

stat_dia_names(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  y_pos = 0.2,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'rect'}

\item{position}{PARAM_DESCRIPTION, Default: 'identity'}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}

\item{lower}{PARAM_DESCRIPTION, Default: 0.25}

\item{upper}{PARAM_DESCRIPTION, Default: 1}

\item{barwidth}{PARAM_DESCRIPTION, Default: 0.9}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
