% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{corrmorant_themes}
\alias{corrmorant_themes}
\alias{theme_corrm}
\title{Themes for ggcorrm plots}
\usage{
theme_corrm(
  base_size = 9,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size (defaults to 9).}

\item{base_family}{Base font family (defaults to "").}

\item{base_line_size}{Base size for line elements (defaults to base_size/22).}

\item{base_rect_size}{Base sie for rect elements (defaults to base_size/22).}
}
\description{
ggplot themes controlling all non-data display meant for the use
with\code{ggcorrm}. Single elements of a theme can be modified with
\code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\details{
Available themes:
\itemize{
\item \code{theme_corrm()}: The standard corrmorant theme with white background and
strong black facet outlines. \emph{So far the only available corrmorant theme.}
}
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}
}
