% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_binned.R
\name{dia_freqpoly}
\alias{dia_freqpoly}
\title{FUNCTION_TITLE}
\usage{
dia_freqpoly(mapping = NULL, lower = 0.25, upper = 1, barwidth = 0.9, ...)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{lower}{PARAM_DESCRIPTION, Default: 0.25}

\item{upper}{PARAM_DESCRIPTION, Default: 1}

\item{barwidth}{PARAM_DESCRIPTION, Default: 0.9}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{character(0)}}
}
