% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_names.R
\name{dia_names}
\alias{dia_names}
\title{FUNCTION_TITLE}
\usage{
dia_names(y_pos = 0.2, mapping = NULL, ...)
}
\arguments{
\item{y_pos}{PARAM_DESCRIPTION, Default: 0.2}

\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{character(0)}}
}
