% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drosera.R
\docType{data}
\name{drosera}
\alias{drosera}
\title{Leaf and petiole size for three African sundew species}
\format{
A data.frame with 150 rows and 6 variables: \describe{
\item{species}{a factor denoting the sundew species (\emph{Drosera capensis},
\emph{madagascariensis} and \emph{venusta})} \item{variety}{a factor denoting the
variety ("typical" and "rubra" for \emph{D. capensis}, "typical" and "alba" for
\emph{D. venusta} and "typical" for \emph{D. madagascariensis})}
\item{petiole_length}{a numeric denoting petiole length in millimeters}
\item{petiole_width}{a numeric denoting petiole width in millimeters}
\item{blade_length}{a numeric denoting leaf blade length in millimeters}
\item{blade_width}{an numeric denoting leaf blade width in millimeters}
}
}
\usage{
drosera
}
\description{
Includes measurements for leaf blade and petiole length and
width for \emph{Drosera capensis} (typical and red form), \emph{Drosera
madagascarienis} (typical form) and \emph{Drosera venusta} (typical and
athocyanin-reduced form).
}
\details{
The \code{drosera} dataset contains 150
observations of leaf and petiole size for five varieties of three African
sundew species: the typical and red forms of \emph{Drosera capensis},  the
typical form of \emph{Drosera madagascarienis} and the typical and
anthocyanin-reduced forms \emph{Drosera venusta}. The measurements were
performed in July 2020 on plants growing on my (R.M. Link's) windowsill in
a matter of a couple of hours with a regular caliper and therefore do not
necessarily live up to high levels of scientific rigor. However, when it
comes to illustrating correlation between biometric variables, the dataset
does its job, and therefore can be used as an replacement for the
\link[datasets:iris]{iris} dataset by Edgar Anderson and Ronald A. Fisher,
which has a problematic past due to its publication in the Annals of
Eugenics.

The \code{drosera} dataset is available as a separate data package under
\url{https://github.com/r-link/drosera}. It
is  published under a Creative Commons Zero license and is therefore free
for all kinds of commercial and non-commercial use. In particular, if you
need a replacement dataset for \code{iris} for the examples in your R package,
feel free to add \code{drosera} to the package without any additional
precautions.

However, it would be kind if you add a link to the package's Github page in
case you decide to use it (and maybe drop me a line so I can link your
project in the repository), but I won't be mad if you don't.
}
\keyword{datasets}
