% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cortext.R
\name{geom_cortext}
\alias{geom_cortext}
\title{FUNCTION_TITLE}
\usage{
geom_cortext(
  mapping = NULL,
  nrow = NULL,
  ncol = NULL,
  digits = 2,
  corrsize = TRUE,
  corr_method = "pearson",
  squeeze = 0.7,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{nrow}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: NULL}

\item{digits}{PARAM_DESCRIPTION, Default: 2}

\item{corrsize}{PARAM_DESCRIPTION, Default: TRUE}

\item{corr_method}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{squeeze}{PARAM_DESCRIPTION, Default: 0.7}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{character(0)}}
}
