% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heatcircle.R
\name{geom_heatcircle}
\alias{geom_heatcircle}
\title{Circles whose size represents correlation strength}
\usage{
geom_heatcircle(
  corr_method = "pearson",
  rmin = 0.1,
  rmax = 0.9,
  scale_by = c("area", "radius"),
  ...
)
}
\arguments{
\item{corr_method}{character string with the correlation method passed
to \code{\link[stats]{cor}}. Can be one of "pearson", "kendall" and
"spearman". Defaults to "pearson" (or is inherited from the setting
in \code{\link[ggcorrm]{ggcorrm()}}).}

\item{rmin}{numeric (0 - 1). Minimum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.1.}

\item{rmax}{numeric (0 - 1). Maximum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.9.}

\item{scale_by}{character.  \code{scale_by = "area"} scales the area of the
circles by correlation strength, \code{scale_by = "radius"} scales by
radius. Defaults to "area".}

\item{...}{Additional arguments to \code{\link{stat_heatcircle}}.}
}
\value{
A \code{ggplot2} layer with circles displaying correlation strength.
}
\description{
\code{geom_heatcircle()} is used to display circles whose fill and
    size aesthetics depoend on the strength of bivariate correlations between
    variables. It is meant for the off-diagonal facets of \code{\link{ggcorrm}}
    plots.
}
\details{
\code{geom_heatcircle()} is a wrapper around
   \code{\link[stat_heatcircle]{stat_heatcircle()}} that additionally takes
   care of the right specification of aesthetics.

   \code{stat_heatcircle()} computes the outlines of circles that
   are centered around the midpoint of \code{gcorrm} facets and whose radius
   or area is modified in dependence of correlation strength. \code{rmin} and
   \code{rmax} allow to modify the range in which the radii of the circles may
   vary.
}
\seealso{
\code{\link{stat_heatcircle}},
  \code{\link{geom_heatmap}}
}
