% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heatmap.R
\name{geom_heatmap}
\alias{geom_heatmap}
\alias{geom_heatpoint}
\title{Correlation heatmaps in ggcorrm plots.}
\usage{
geom_heatmap(corr_method = "pearson", ...)

geom_heatpoint(corr_size = TRUE, corr_method = "pearson", ...)
}
\arguments{
\item{corr_method}{character string with the correlation method passed
to \code{\link[stats]{cor}}. Can be one of "pearson", "kendall" and
"spearman". Defaults to "pearson" (or is inherited from the setting
in \code{\link[ggcorrm]{ggcorrm()}}).}

\item{...}{Additional arguments to \code{\link{stat_heatmap}}.}

\item{corr_size}{logical - should the \code{size} aesthetic of the
points in `geom_heatpoint()` be expressed as a function of correlation
strength? \code{corr_size = TRUE} Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot2} layer with correlation heatmaps / points indicating
   correlation strength.
}
\description{
\code{geom_heatmap()} is used to display correlation heatmaps
    in the off-diagonal facets of \code{\link{ggcorrm}} plots.
    \code{geom_heatpoint()} scales the and colour of a single centered
    \code{\link[ggplot2]{geom_point}} by correlation strength.
}
\details{
\code{geom_heatmap()} can be used to display correlation heat maps
   in the facets of \code{ggcorrm} plots.

   \code{geom_heatmap()} is a wrapper around
   \code{\link[stat_heatmap]{stat_heatmap()}} that additionally takes
   care of the right specification of aesthetics.

   \code{geom_heatpoint()} instead plots a single point in the middle of the
   layer whose size on default depends on the strength of correlation. The
   shape parameter as well as transparency etc. can be adjusted via the
   \code{...} argument.
   The range of sizes covered by the 'heatpoints' can be adjusted by
   \code{\link[ggplot2:scale_size]{scale_size()}}.

   For larger circles that scale with correlation, see
   \code{\link{geom_heatcircle()}}.
}
\seealso{
\code{\link{stat_heatmap}},
  \code{\link{geom_heatcircle}}
}
