% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cortext.R
\name{get_cortext_pos}
\alias{get_cortext_pos}
\title{FUNCTION_TITLE}
\usage{
get_cortext_pos(stats, nrow = NULL, ncol = NULL, squeeze, xrange, yrange)
}
\arguments{
\item{stats}{PARAM_DESCRIPTION}

\item{nrow}{PARAM_DESCRIPTION, Default: NULL}

\item{ncol}{PARAM_DESCRIPTION, Default: NULL}

\item{squeeze}{PARAM_DESCRIPTION}

\item{xrange}{PARAM_DESCRIPTION}

\item{yrange}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
