% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrm.R
\name{ggcorrm}
\alias{ggcorrm}
\title{FUNCTION_TITLE}
\usage{
ggcorrm(
  data,
  labels = NULL,
  rescale = c("by_sd", "by_max", NULL),
  corr_group = NULL,
  corr_method = "pearson",
  mutates = NULL,
  bg_dia = NULL,
  bg_lotri = NULL,
  bg_utri = NULL
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{labels}{PARAM_DESCRIPTION, Default: NULL}

\item{rescale}{PARAM_DESCRIPTION, Default: c("by_sd", "by_max", NULL)}

\item{corr_group}{PARAM_DESCRIPTION, Default: NULL}

\item{corr_method}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{mutates}{PARAM_DESCRIPTION, Default: NULL}

\item{bg_dia}{PARAM_DESCRIPTION, Default: NULL}

\item{bg_lotri}{PARAM_DESCRIPTION, Default: NULL}

\item{bg_utri}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}
}
