% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.infinite.data.frame}
\alias{is.infinite.data.frame}
\title{FUNCTION_TITLE}
\usage{
\method{is.infinite}{data.frame}(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
