% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.waive}
\alias{is.waive}
\title{FUNCTION_TITLE}
\usage{
is.waive(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
