% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_selectors.R
\name{lotri}
\alias{lotri}
\title{FUNCTION_TITLE}
\usage{
lotri(geom)
}
\arguments{
\item{geom}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
