% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_list}
\alias{modify_list}
\title{FUNCTION_TITLE}
\usage{
modify_list(old, new)
}
\arguments{
\item{old}{PARAM_DESCRIPTION}

\item{new}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
