% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_var}
\alias{rescale_var}
\title{FUNCTION_TITLE}
\usage{
rescale_var(x, lower, upper, range, append_x = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lower}{PARAM_DESCRIPTION}

\item{upper}{PARAM_DESCRIPTION}

\item{range}{PARAM_DESCRIPTION}

\item{append_x}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
