% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_corr}
\alias{scale_fill_corr}
\title{FUNCTION_TITLE}
\usage{
scale_fill_corr(
  ...,
  option = c("A", "B", "C", "D"),
  limits = c(-1, 1),
  na.value = "#55FF55",
  guide = "colourbar",
  aesthetics = "fill",
  direction = 1
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{option}{PARAM_DESCRIPTION, Default: c("A", "B", "C", "D")}

\item{limits}{PARAM_DESCRIPTION, Default: c(-1, 1)}

\item{na.value}{PARAM_DESCRIPTION, Default: '#55FF55'}

\item{guide}{PARAM_DESCRIPTION, Default: 'colourbar'}

\item{aesthetics}{PARAM_DESCRIPTION, Default: 'fill'}

\item{direction}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
