% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skew}
\alias{skew}
\title{FUNCTION_TITLE}
\usage{
skew(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{na.fail}}
}
