% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cortext.R
\name{stat_cortext}
\alias{stat_cortext}
\title{FUNCTION_TITLE}
\usage{
stat_cortext(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  nrow = nrow,
  ncol = ncol,
  digits = 2,
  corr_method = "pearson",
  squeeze = 0.7,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'text'}

\item{position}{PARAM_DESCRIPTION, Default: 'identity'}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}

\item{nrow}{PARAM_DESCRIPTION, Default: nrow}

\item{ncol}{PARAM_DESCRIPTION, Default: ncol}

\item{digits}{PARAM_DESCRIPTION, Default: 2}

\item{corr_method}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{squeeze}{PARAM_DESCRIPTION, Default: 0.7}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
