% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dia_freqpoly.R, R/geom_dia_histogram.R,
%   R/stat_dia_bin.R
\name{geom_dia_freqpoly}
\alias{geom_dia_freqpoly}
\alias{geom_dia_histogram}
\alias{stat_dia_bin}
\title{Compute histograms and frequency polygons for ggcorrm plots.}
\usage{
geom_dia_freqpoly(
  mapping = NULL,
  data = NULL,
  stat = "dia_bin",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = FALSE
)

geom_dia_histogram(
  mapping = NULL,
  data = NULL,
  stat = "dia_bin",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = FALSE
)

stat_dia_bin(
  mapping = NULL,
  data = NULL,
  geom = "dia_histogram",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  bins = 10,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Additional arguments passed \code{\link[ggplot2:layer]{ggplot2::layer()}} (arguments for
\link[ggplot2:geom_histogram]{ggplot2::stat_bin()} are permitted).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom}{The geometric object to use display the data}

\item{lower}{numeric between 0 and 1. Lower limit of the histograms/frequency
polygons relative to the range of the \code{y} axis. Defaults to 0.25.}

\item{upper}{numeric between 0 and 1. Upper limit of the histograms/frequency
polygons relative to the range of the \code{y} axis. Defaults to 1.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_dia_bin()} computes the binned data summaries for the
diagonal panels of \link{ggcorrm} plots that are created with \link{dia_histogram}
and \link{dia_freqpoly}.
}
\details{
\code{stat_dia_bin()} computes binned data summaries for display in the
diagonal facets of \code{ggcorrm} plots. The \code{lower} and \code{upper} arguments can
be used to offset the histograms/frequency polygons from zero and optimally
fit them to the range of each panel.
}
\seealso{
\preformatted{[ggplot2::stat_bin()][ggplot2::geom_histogram()],
}

\code{\link[=dia_histogram]{dia_histogram()}},
\code{\link[=dia_freqpoly]{dia_freqpoly()}}
}
