% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funtext.R
\name{stat_funtext}
\alias{stat_funtext}
\title{Compute text labels based on user-specified functions}
\usage{
stat_funtext(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  fun,
  nrow = NULL,
  ncol = NULL,
  squeeze = 0.7,
  byrow = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{fun}{One of the following: * a) a function of \code{x} and \code{y}. * b) an
rlang lambda style \link[rlang:as_function]{one-sided formula} describing a
function of two parameters labeled\code{.x} and \code{.y} * c) an rlang quosure
created with \code{quo()} that is evaluated in the context of the data analogous
to the \code{mutates} argument in \code{\link[=tidy_corrm]{tidy_corrm()}}.

The functions in a) and b) are evaluated using the \code{x} and \code{y}coordinates
of the plot as first and second argument, respectively. The quosure in c)
is evaluated in the context of the raw data and may contain x and y.

One outcome of the computation per group is stored in a new column called
\code{fun_out} that can be accessed in the aesthetics using \code{after_stat(fun_out)}.}

\item{nrow}{integer - the number of rows for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{nrow}).}

\item{ncol}{integer - the number of columns for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{ncol}).}

\item{squeeze}{numeric between 0 an 1. Proportion of the facet width/height
the facet labels are restricted to when multiple labels are present
(defaults to 0.7 - labels extend over 70\% of the extent of the plot).}

\item{byrow}{logical. Should the correlation labels in plots with multiple
groups be filled by rows (\code{byrow = TRUE}) or by columns (\code{byrow = FALSE})?
Note that the actual number of rows or columns that are filled with values
can be below the specified value of \code{nrow} or \code{ncol} when less rows/columns
than specified are needed to reach the total number of groups. Defaults to
\code{TRUE}.}

\item{...}{additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_funtext()} is used to compute
the facets of \link{ggcorrm} plots.
}
\details{
\code{stat_funtext()} evaluates a function, rlang lambda style \link[rlang:as_function]{one-sided formula} or a quosure in the context of the data used
for plotting and computes positions in the plot using the same
\code{nrow}/\code{ncol} based positioning as \code{\link[=stat_corrtext]{stat_corrtext()}}.
}
\seealso{
\code{\link[=stat_corrtext]{stat_corrtext()}} for correlation text labels.
\code{\link[=lotri_funtext]{lotri_funtext()}} and \code{\link[=utri_funtext]{utri_funtext()}}
}
