% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heatcircle.R
\name{stat_heatcircle}
\alias{stat_heatcircle}
\title{Compute circles with correlation-dependent size}
\usage{
stat_heatcircle(
  mapping = NULL,
  data = NULL,
  geom = GeomHeatcircle,
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  corr_method = NULL,
  rmin = 0.1,
  rmax = 0.9,
  scale_by = c("area", "radius"),
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{corr_method}{character string with the correlation method passed to
\link[stats:cor]{cor()}. Can be one of \code{"pearson"}, \code{"kendall"} and
\code{"spearman"}. Defaults to \code{"pearson"} (or is inherited from the setting in
\code{\link[=ggcorrm]{ggcorrm()}}).}

\item{rmin}{numeric (0 - 1). Minimum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.1.}

\item{rmax}{numeric (0 - 1). Maximum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.9.}

\item{scale_by}{character.  \code{scale_by = "area"} scales the area of the
circles by correlation strength, \code{scale_by = "radius"} scales by radius.
Defaults to "area".}

\item{...}{additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_heatcircle()} is used to compute bivariate correlations
and display them in \link{ggcorrm} plots in form of circles whose color and size
indicate correlation strength.
}
\details{
\code{stat_heatcircle()} computes the outlines of circles that are
centered around the midpoint of \code{gcorrm} facets and whose radius or area is
modified in dependence of correlation strength. \code{rmin} and \code{rmax} allow to
modify the range in which the radii of the circles may vary. The circles
are displayed using \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}},
\code{\link[=add_heatcircle]{add_heatcircle()}}
}
