% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_corrm}
\alias{theme_corrm}
\title{FUNCTION_TITLE}
\usage{
theme_corrm(
  base_size = 9,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{PARAM_DESCRIPTION, Default: 9}

\item{base_family}{PARAM_DESCRIPTION, Default: ''}

\item{base_line_size}{PARAM_DESCRIPTION, Default: base_size/22}

\item{base_rect_size}{PARAM_DESCRIPTION, Default: base_size/22}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grid]{unit}}
}
