% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_corrm.R
\name{tidy_corrm}
\alias{tidy_corrm}
\title{FUNCTION_TITLE}
\usage{
tidy_corrm(
  data,
  labels = NULL,
  rescale = c("by_sd", "by_range", NULL),
  corr_group = NULL,
  corr_method = "pearson",
  mutates = NULL
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{labels}{PARAM_DESCRIPTION, Default: NULL}

\item{rescale}{PARAM_DESCRIPTION, Default: c("by_sd", "by_range", NULL)}

\item{corr_group}{PARAM_DESCRIPTION, Default: NULL}

\item{corr_method}{PARAM_DESCRIPTION, Default: 'pearson'}

\item{mutates}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_nest}},\code{\link[dplyr]{select}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{case_when}}
 \code{\link[stats]{sd}}
 \code{\link[tidyr]{gather}},\code{\link[tidyr]{expand_grid}},\code{\link[tidyr]{nest}}
 \code{\link[purrr]{map}}
}
