% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_selectors.R
\name{update_data}
\alias{update_data}
\title{FUNCTION_TITLE}
\usage{
update_data(data, pos)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{pos}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{character(0)}}
 \code{\link[methods]{is}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{join}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}}
 \code{\link[tidyr]{nest}}
}
