% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funtext.R
\name{add_funtext}
\alias{add_funtext}
\alias{lotri_funtext}
\alias{utri_funtext}
\title{User-specified text labels in off-diagonal ggcorrm facets}
\usage{
lotri_funtext(
  fun,
  mapping = NULL,
  nrow = NULL,
  ncol = NULL,
  squeeze = 0.7,
  byrow = TRUE,
  show.legend = FALSE,
  ...
)

utri_funtext(
  fun,
  mapping = NULL,
  nrow = NULL,
  ncol = NULL,
  squeeze = 0.7,
  byrow = TRUE,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \link[ggplot2:aes]{aes()}. x
and y are set automatically and must not bechanged, but all other
aesthetics may be manipulated. By default, the \code{fill} aesthetic is mapped
to \code{..fun_out..} internally, but this is overridden when explicitly
specified. Defaults to \code{NULL} (use standard settings).}

\item{nrow}{integer - the number of rows for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{nrow}).}

\item{ncol}{integer - the number of columns for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{ncol}).}

\item{squeeze}{numeric between 0 an 1. Proportion of the facet width/height
the facet labels are restricted to when multiple labels are present
(defaults to 0.7 - labels extend over 70\% of the extent of the plot).}

\item{byrow}{logical. Should the correlation labels in plots with multiple
groups be filled by rows (\code{byrow = TRUE}) or by columns (\code{byrow = FALSE})?
Note that the actual number of rows or columns that are filled with values
can be below the specified value of \code{nrow} or \code{ncol} when less rows/columns
than specified are needed to reach the total number of groups. Defaults to
\code{TRUE}.}

\item{show.legend}{logical. Should this layer be included in the legends?
FALSE (the default) never includes, TRUE always includes, and NA includes
only if aesthetics are mapped. It can also be a named logical vector
to finely select the aesthetics to display.}

\item{...}{Additional arguments to \link{stat_funtext}.}
}
\value{
A \code{ggplot2} layer with text labels for user_specified functions.
}
\description{
\code{lotri_funtext()} and \code{utri_funtext()} are used to display
user-specified text labels in the off-diagonal facets of \link{ggcorrm} plots.
}
\details{
\code{lotri_funtext()} and \code{utri_funtext()} can be used to display user-
defined text labels in the lower or upper triangular facets of
\code{ggcorrm} plots, respectively. The functions can be evaluated separately
for different groups by using grouping aesthetics such as \code{aes(color = group)}.

If \code{fun} is a function or rlang lambda-style \link[rlang:as_function]{one-sided formula},
it is called on the data in each group using the \code{x} coordinate as first and the \code{y}
coordinate as second argument. If \code{fun} is a quosure, it is evaluated in the
context of the data in the respective group.

\code{lotri_funtext()} and \code{utri_funtext()} are wrappers around
\code{\link[=stat_funtext]{stat_funtext()}} that additionally take care of the right specification of
aesthetics. The positioning of the labels in \code{\link[=stat_funtext]{stat_funtext()}} follows the
same rules as in \code{\link[=stat_corrtext]{stat_corrtext()}}.
}
\examples{
\dontrun{
if(interactive()){

# function to compute linear model slopes
lmslope <- function(y, x)  round(coef(lm(y ~ x))[2], 2)

# add slopes using a function
ggcorrm(drosera, rescale = "as_is") +
  lotri(geom_point(alpha = 0.4)) +
  utri_funtext(fun = lmslope) +
  dia_density(fill = "steelblue", lower = .4) +
  dia_names(y_pos = .1)

# compute Pearson correlations as a lambda expression (top)
# and as a quosure(bottom)
ggcorrm(drosera, rescale = "as_is") +
  utri_funtext(fun = ~ round(cor(.x, .y), 3)) +
  lotri_funtext(fun = quo(round(cor(x, y), 3))) +
  dia_names(y_pos = .5)

}
}

}
\seealso{
\link{stat_funtext}
}
