% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant.R
\name{corrmorant}
\alias{corrmorant}
\title{FUNCTION_TITLE}
\usage{
corrmorant(
  data,
  style = c("dark", "light", "blue_red"),
  rescale = c("by_sd", "by_range", NULL),
  labels = NULL,
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{style}{PARAM_DESCRIPTION, Default: c("dark", "light", "blue_red")}

\item{rescale}{PARAM_DESCRIPTION, Default: c("by_sd", "by_range", NULL)}

\item{labels}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
