% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant.R
\name{corrmorant}
\alias{corrmorant}
\title{Simple automated correlation plots}
\usage{
corrmorant(
  data,
  style = c("blue_red", "dark", "light", "binned"),
  rescale = c("as_is", "by_sd", "by_range"),
  corr_method = c("pearson", "kendall", "spearman"),
  labels = NULL,
  bins = 10,
  ...
)
}
\arguments{
\item{data}{For \code{tidy_corrm.default}, a \code{data.frame} or \code{matrix} with the
raw data used for the correlation plot. If a \code{data.frame}, all numeric
variables are used as rows/columns of the correlation plot, while all other
variables are appended to the reshaped dataset as additional columns.}

\item{style}{Character string defining the plot style. One of \code{"dark"},
\code{"light"}, \code{"blue_red"} or \code{"binned"}.}

\item{rescale}{character string specifying the type of transformation
performed on the numeric variables in the plot. The standard argument
\verb{as_is"} uses the unchanged raw values.\code{"by.sd"} scales by the standard
deviation of the data and centers around zero.  \code{"by.range"} rescales the
range of the data to the interval from 0 to 1.  Defaults to  \code{"as_is"}.}

\item{corr_method}{character string with the correlation method passed to
\code{\link[stats:cor]{stats::cor()}}. Used for the \code{.corr} variable appended to the \code{tidy_corr}
dataset and passed on to \code{\link[=lotri_corrtext]{lotri_corrtext()}}/ \code{\link[=utri_corrtext]{utri_corrtext()}}  layers. Can
be one of \code{"pearson"}, \code{"kendall"}  and \code{"spearman"}. Defaults to
\code{"pearson"}.}

\item{labels}{(Optional) character vector with labels for the names of all
numeric columns that are used to replace the column names in the plot axis
and text labels. Must be of the same length as the number of numeric
columns displayed in the plot. Defaults to \code{NULL} (use original column
names as labels).}

\item{bins}{Integer. Number of bins for \code{style = "binned"} (ignored if using
other styles). Defaults to 10.}

\item{...}{Additional arguments to \code{\link[=ggcorrm]{ggcorrm()}}.}
}
\description{
\code{corrmorant()} is a wrapper around \code{\link[=ggcorrm]{ggcorrm()}} that creates
scatterplot matrices with correlation output using reasonable standard
values for initial data inspection.
}
\details{
\code{corrmorant()} is a simplified wrapper around \code{\link[=ggcorrm]{ggcorrm()}} that
creates scatterplot matrices with reasonable standard settings. Setting the
\code{style} argument to \code{"dark"}, \code{"light"} or \code{"blue_red"} produces regular
scatterplots with different color schemes. \code{style = "binned"} replicates
\code{style = "blue_red"} with a binned scatterplot (the number of bins defaults
to 10, but can be changed with the \code{bins} argument), which can be very
useful to reduce the plotting time for large datasets. Refer to the
documentation of \code{\link[=ggcorrm]{ggcorrm()}} and \code{\link[=tidy_corrm]{tidy_corrm()}} for more details.
}
\examples{
\dontrun{
if(interactive()){
 # plot the drosera dataset with 3 different styles
 corrmorant(drosera, style = "dark")
 corrmorant(drosera, style = "light")
 corrmorant(drosera, style = "blue_red")
 }
}
}
\seealso{
\link{ggcorrm}, \link{tidy_corrm}
}
