% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmorant_ggproto.R, R/geom_dia_density.R,
%   R/geom_dia_freqpoly.R, R/geom_dia_histogram.R, R/geom_reltext.R,
%   R/stat_corrtext.R, R/stat_dia_bin.R, R/stat_dia_names.R, R/stat_funtext.R,
%   R/stat_heatcircle.R, R/stat_heatmap.R, R/tidy_corrm.R
\docType{data}
\name{corrmorant_ggproto}
\alias{corrmorant_ggproto}
\alias{GeomDiaDensity}
\alias{GeomDiaFreqpoly}
\alias{GeomDiaHistogram}
\alias{GeomReltext}
\alias{StatCorrtext}
\alias{StatDiaBin}
\alias{StatDiaNames}
\alias{StatFuntext}
\alias{StatHeatcircle}
\alias{StatHeatmap}
\alias{TidyCorrm}
\title{ggproto classes for corrmorant}
\format{
An object of class \code{TidyCorrm} (inherits from \code{ggproto}, \code{gg}) of length 4.
}
\usage{
TidyCorrm
}
\description{
Class definitions for the new \code{ggproto} classes defined in the
\code{corrmorant} package.
}
\section{New stats}{

\itemize{
\item \code{TidyCorrm}: prototype for \code{\link[=tidy_corrm]{tidy_corrm()}} constructor methods.
\item \code{StatCorrtext}: create and position text labels for correlation strength
(see \link{stat_corrtext} for details).
\item \code{StatCorrtext}: create and position text labels for the outcome of user
defined functions (see \link{stat_funtext} for details).
\item \code{StatHeatmap}: create correlation heat maps (see \link{stat_heatmap} for
details).
\item \code{StatHeatcircle}: create circles indicating correlation strength (see
\link{stat_heatcircle} for details).
\item \code{StatDiaNames}: create and position text labels for variable names in the
diagonal facets of \code{ggcorrm} plots (see \link{stat_dia_names} for details).
\item \code{StatDiaBin}: create and position plots of binned data (histograms or
frequency polygons) of variables in the diagonal facets of \code{ggcorrm} plots
(see \link{stat_dia_bin} for details).
}
}

\section{New geoms}{

\itemize{
\item \code{GeomReltext}: text labels with relative positions (see \link{geom_reltext} for
details).
\item \code{GeomDiaDensity}: create and position density plots in the diagonal facets
of \code{ggcorrm} plots (see \link{geom_dia_density} for details).
\item \code{GeomDiaHistogram}: create and position histogramsin the diagonal facets
of \code{ggcorrm} plots (see \link{geom_dia_density} for details).
\item \code{GeomDiaFreqpoly}: create and position frequency polygons  in the diagonal
facets of \code{ggcorrm} plots (see \link{geom_dia_density} for details).
}
}

\seealso{
\itemize{
\item \link{ggplot2-ggproto} for a description of the \code{ggproto} object orientation
system,
\item \link{tidy_corrm} for structured tidy correlation plot datasets,
\item \link{add_corrtext} for correlation text labels,
\item \link{add_funtext} for user-defined text labels,
\item \link{add_heatmap} for correlation heatmaps,
\item \link{add_heatcircle} for circles witch correlation-dependent size and color,
\item \link{dia_names} for variable names in the plot diagonal,
\item \link{dia_density} for density plots in the plot diagonal,
\item \link{dia_histogram} for histograms and frequency polygons in the plot diagonal
}
}
\keyword{datasets}
