% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_bin.R
\name{dia_histogram}
\alias{dia_histogram}
\title{FUNCTION_TITLE}
\usage{
dia_histogram(
  mapping = NULL,
  lower = 0.25,
  upper = 1,
  barwidth = 0.9,
  position = "dodge",
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{lower}{PARAM_DESCRIPTION, Default: 0.25}

\item{upper}{PARAM_DESCRIPTION, Default: 1}

\item{barwidth}{PARAM_DESCRIPTION, Default: 0.9}

\item{position}{PARAM_DESCRIPTION, Default: 'dodge'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
