% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_bin.R
\name{dia_histogram}
\alias{dia_histogram}
\alias{dia_freqpoly}
\title{Histograms and frequency polygons for ggcorrm plots}
\usage{
dia_histogram(
  mapping = NULL,
  lower = 0.25,
  upper = 1,
  barwidth = 0.9,
  position = "dodge",
  ...
)

dia_freqpoly(mapping = NULL, lower = 0.25, upper = 1, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2:aes]{aes()}}. \code{x} and \code{y} are set
automatically and must not be changed, but all other aesthetics
may be manipulated.}

\item{lower}{numeric between 0 and 1. Lower limit of the
histograms/frequency polygons relative to the range of the \code{y}
axis. Defaults to 0.25.}

\item{upper}{numeric between 0 and 1. Upper limit of the
histograms/frequency polygons relative to the range of the \code{y}
axis. Defaults to 1.}

\item{barwidth}{Width of the histograms relative to the maximum possible
width. Defaults to 0.9.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Additional parameters for \code{\link{stat_dia_bin}}.}
}
\value{
A \code{ggplot2} layer with histograms or frequency polygons for the
   variables on the plot diagonal of \code{ggcorrm} plots.#'
}
\description{
Add histograms or frequency polygons to the diagonal panels of
     \code{\link{ggcorrm}} plots.
}
\details{
The \code{lower} and\code{upper} arguments can be used to offset
    the histograms/frequency polygons from zero and optimally fit them to
    the range of each panel.
    The standard values are chosen to work well when placing text labels
    under the histograms/frequency polygons with \code{\link{dia_names}}.

   \code{dia_histogram()} adds histograms of the numeric variables in a
   \code{ggcorrm} plot to the plot diagonal. The bar width can be adjusted
   with \code{barwidth}. Frequency polygons can be created with
   \code{dia_freqpoly()}. Both functions use the same stat,
   \code{\link{stat_dia_bin}}.
}
\seealso{
\code{\link[ggplot2:stat_bin]{ggplot2::stat_bin}},
    \code{\link{stat_dia_bin}}
}
