% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_names.R
\name{stat_dia_names}
\alias{stat_dia_names}
\alias{dia_names}
\title{Add variable names to diagonal facets}
\usage{
stat_dia_names(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  y_pos = 0.2,
  ...
)

dia_names(y_pos = 0.2, mapping = NULL, ...)
}
\arguments{
\item{mapping}{(optional) mapping for the internal
\code{\link[ggplot2:geom_text]{geom_text}()} call. Must be created with
\code{\link[ggplot2:aes]{aes}()}, and must not contain \code{x} or
\code{y}. Can e.g. be used to color text labels by groups (see example).
Defaults to NULL.}

\item{y_pos}{numeric between 0 and 1 specifying the relative position of the
text labels along the x axis (defaults to 0.2).}

\item{...}{Additional parameters passed to
\code{\link[ggplot2:geom_text]{geom_text}()}.}
}
\value{
A layer containing text labels on the diagonal facets.
}
\description{
\code{dia_names()} is a wrapper around \code{stat_dia_names()}
   which plots the names of variables in a \code{ggcorrm} plot at the
   required position in the diagonal facets.
}
\details{
\code{dia_names()} plots text labels on the plot diagonal and takes
   care of the often complicated positioning in plots with different x and y
   ranges.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2:geom_text]{geom_text}()},
  \code{\link{ggcorrm}},
  \code{\link{tidy_corrm}},
  \code{\link{dia_density}},
  \code{\link{dia_histogram}},
  \code{\link{dia_freqpoly}}
}
