% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_reltext.R
\name{geom_reltext}
\alias{geom_reltext}
\title{Add text to relative positions}
\usage{
geom_reltext(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \link[ggplot2:aes]{aes} that
are passed on to subsequent layers. \code{x} and \code{y} are set automatically and
must not be changed,  but all other aesthetics may be manipulated. Defaults
to \code{NULL} (use standard \code{ggcorrm} mapping).}

\item{data}{Dataset used to compute the correlation matrix. Can be either a
\code{data.frame} or \code{matrix}, or an object of class \code{tidy_corrm}. If specifying
a \code{data.frame} or \code{matrix}, it will internally be passed to \code{\link[=tidy_corrm]{tidy_corrm()}}
with the settings for \code{labels}, \code{rescale}, \code{corr_group}, \code{corr_method} and
\code{mutates} specified in the \code{ggcorrm()} call. \code{\link[=tidy_corrm]{tidy_corrm()}} prepares the
data for plotting by creating a \code{data.frame} with all possible combinations
of all numeric variables, while retaining all discrete variables in
additional columns. If a \code{tidy_corrm} object is supplied, all arguments
passed to \code{\link[=tidy_corrm]{tidy_corrm()}} will be ignored and the \code{tidy_corrm} object wil be
used directly to initialize the \code{ggcorrm} plot.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{geom_reltext()} adds text labels to relative positions in a
ggplot panel.
}
\details{
\code{geom_reltext()} adds text to relative positions in ggplot panels,
i.e. the position of the text is expressed as a number between 0 and 1
(lower and upper axis limits) instead of a number on the scale of the
displayed variables.

In corrmorant, it is widely used to display text labels relative to the
panel margins instead of the data in the panel.
}
\section{Aesthetics}{

\code{geom_reltext()} requires the following aesthetics:
\itemize{
\item \strong{relx} (relative position on the x-axis)
\item \strong{rely} (relative position on the y-axis)
\item \strong{label}
}
In addition, it understands the same aesthetics as \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}
}

\seealso{
Used to position the output of \code{\link[=stat_corrtext]{stat_corrtext()}}, \code{\link[=stat_corrtext]{stat_corrtext()}} and
\code{\link[=stat_dia_names]{stat_dia_names()}}.
}
