% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrm.R
\name{ggcorrm}
\alias{ggcorrm}
\title{Create a ggcorrm correlation plot}
\usage{
ggcorrm(
  data,
  labels = NULL,
  rescale = c("by_sd", "by_max", NULL),
  corr_method = "pearson",
  corr_group = NULL,
  mutates = NULL,
  bg_dia = NULL,
  bg_lotri = NULL,
  bg_utri = NULL
)
}
\arguments{
\item{data}{Dataset used to compute the correlation matrix. Can be either a
\code{data.frame} or \code{matrix}, or an object of class \code{tidy_corrm}.
If specifying a \code{data.frame} or \code{matrix}, it will internally be
passed to \code{\link{tidy_corrm()}} with the settings for \code{labels},
\code{rescale}, \code{corr_group}, \code{corr_method} and \code{mutates}
specified in the \code{ggcorrm()} call. \code{\link{tidy_corrm()}}
prepares the dataset for plotting by creating a \code{data.frame} with
all possible combinations of all numeric variables while retaining all
discrete variables in additional columns. If a \code{tidy_corrm} object
is supplied, all arguments passed to \code{\link{tidy_corrm()}} will be
ignored and the \code{tidy_corrm} object wil be used directly to
initialize the \code{ggcorrm} plot.}

\item{labels}{(optional) character vector with labels for the names of all
numeric columns that are used to replace the column names in the plot
axis and text labels. Must be of the same length as the number of
numeric columns displayed in the plot. Default: \code{NULL}}

\item{rescale}{a character string specifying the type of transformation
performed on the numeric variables in the plot. The standard argument
\code{"by.sd"} scales by the standard deviation of the data and centers
around zero.  \code{"by.range"} rescales the range of the data to the
interval from 0 to 1. Use \code{rescale = NULL} to use the unchanged raw
values. Default: \code{"by_sd"}}

\item{corr_method}{character string with the correlation method passed to
\code{\link[stats]{cor}}. Can be of \code{"pearson"}, \code{"kendall"}
 and \code{"spearman"}
 Default: \code{"pearson"}}

\item{corr_group}{NULL or the name of a numeric variable in \code{data}. If
a grouping variable is specified, \code{.corr} which can be used for
conditional coloring will be calculated separately for each of these
groups.
Default: \code{NULL}}

\item{mutates}{(optional) list of named quosures created with
\code{\link[rlang:quos]{rlang::quos}}. Can be any expressions that specify
 changes to the `tidy_corrm` \emph{after} reshaping, using regular
\code{\link[dplyr:mutate]{dplyr::mutate}} syntax. Default: NULL}

\item{bg_dia}{(optional) background color specification for the diagonal panels.
Either a character string with a hexadecimal color code, a character string
specifying a color name in \code{\link[grDevices]{colors}}, or an integer
specifying a position in \code{\link[grDevices]{palette}}. The default value
of \code{NULL} used the standard background color defined in the corresponding
\code{\link[ggplot2:theme]{ggplot2::theme}}.}

\item{bg_lotri}{(optional) background color specification for the panels in the
lower triangle. Either a character string with a hexadecimal color code, a
character string specifying a color name in \code{\link[grDevices]{colors}},
or an integer specifying a position in \code{\link[grDevices]{palette}}. The
default value of \code{NULL} used the standard background color defined in the
corresponding \code{\link[ggplot2:theme]{ggplot2::theme}}.}

\item{bg_utri}{(optional) background color specification for the panels in the
lower triangle. Either a character string with a hexadecimal color code, a
character string specifying a color name in \code{\link[grDevices]{colors}},
or an integer specifying a position in \code{\link[grDevices]{palette}}. The
default value of \code{NULL} used the standard background color defined in the
corresponding \code{\link[ggplot2:theme]{ggplot2::theme}}.}
}
\value{
An object of class \code{ggcorrm} containing the reshaped dataset for the
    correlation plot and an empty \code{ggplot} object with appropriate facet and
    theme specifications.
}
\description{
\code{ggcorrm()} initializes a \code{ggcorrm} object (inheriting
    from class \code{ggplot}). It can be called either with a raw data frame
    with input data, which is then internally passed to
    \code{\link{tidy_corrm()}}, or with a prepared correlation object of class
    \code{tidy_corrm}.
}
\details{
\code{ggcorrm} creates the initial correlation plot object containing
    information about panel placement, correlations, themes etc. Its output is a
    modified empty \code{ggplot} object with appropriate facet and theme
    specifications. If a \code{tidy_corrm} object is supplied as \code{data},
    it will be directly plotted without invoking \code{link{tidy_corrm}}, else,
    \code{ggcorrm} passes the raw data and additional arguments to
    \code{link{tidy_corrm}} before plotting (see documentation of this function
    for details).

    New layers can be added using classical
    \code{\link[ggplot:ggplot_add]{ggplot::ggplot_add}}
    syntax, though in most cases it will be more useful to add layers using the
    \code{\link[corrmorant_selectors]{corrmorant selector functions}} which allow
    to map geoms to a subset of panels on the plot diagonal, lower or upper
    triangle with \code{dia()}, \code{lotri()} and \code{utri()}, respectively
    (see examples).

    \code{bg_dia}, \code{bg_lotri} and \code{bg_utri} allow to specify different
    background color settings for the plot diagonal, the lower and the upper
    triangle of the correlation plot matrix, respectively. All other graphics
    settings can be modified using regular ggplot2 \code{\link[ggplot2]{theme}}
    syntax, building upon the ggcorrm standard theme (\code{\link{theme_corrm}}).
}
\examples{
\dontrun{
if(interactive()){
# correlation matrix for the iris dataset
ggcorrm(iris, bg_dia = "grey20") +
  lotri(geom_point(alpha = 0.4)) +
  utri(geom_cortext()) +
  dia_histogram(lower = .3, upper = 1, fill = "grey90", col = 1) +
  dia_names(y_pos = .1, col = "white", size = 3)

# iris data with conditional coloring by pearson correlation
ggcorrm(iris, rescale = "by_sd") +
  lotri(geom_point(aes(col = .corr), alpha = 0.6)) +
  lotri(geom_smooth(aes(fill = .corr, col = .corr),
                    method = "lm", size = 0.3, alpha = 0.6)) +
  utri(geom_cortext(aes(color = .corr))) +
  dia_density(fill = "grey80", lower = .4) +
  dia_names(y_pos = .1) +
  scale_color_corr(option = "A", aesthetics = c("fill", "color"))

# correlation separated by groups
ggcorrm(iris, rescale = "by_sd", bg_dia = "grey95") +
  lotri(geom_point(aes(col = Species), alpha = 0.4)) +
  lotri(geom_smooth(aes(x, y, fill = Species), col = 1, method = "lm"))  +
  utri(geom_cortext(aes(col = Species), nrow = 2)) +
  dia_density(aes(fill = Species), alpha = 0.5, lower = 0.4) +
  dia_names(y_pos = 0.15)

# using the 'mutates' argument to color diagonal panels by leaf type
ggcorrm(iris, rescale = "by_sd",
        mutates = quos(leaftype = ifelse(substr(var_x, 1, 1) == "S",
                                         "Sepal", "Petal"))) +
  lotri(geom_point(alpha = 0.4))+
  utri(geom_cortext()) +
  dia_density(lower = .3,
              mapping = aes(fill = leaftype)) +
  dia_names(y_pos = .1,
            mapping = aes(col = leaftype))
 }
}
}
\seealso{
\code{\link{tidy_corrm}},
 \code{\link{corrmorant}},
 \code{\link[ggplot2:ggplot]{ggplot2::ggplot}},
 \code{\link[ggplot2:theme]{ggplot2::theme}},
 \code{\link[dplyr:mutate]{dplyr::mutate}},
 \code{\link[rlang:quos]{rlang::quos}}
}
