% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_corrtext.R
\name{stat_corrtext}
\alias{stat_corrtext}
\title{Compute correlation strength for text labels in ggcorrm plots}
\usage{
stat_corrtext(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  nrow = NULL,
  ncol = NULL,
  digits = 2,
  corr_method = NULL,
  squeeze = 0.7,
  byrow = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \link[ggplot2:aes]{aes}. \code{x}
and \code{y} are set automatically and must not be changed,  but all other
aesthetics may be manipulated. Defaults to \code{NULL} (use standard \code{ggcorrm}
mapping).}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
FALSE (the default) never includes, TRUE always includes, and NA includes
only if aesthetics are mapped. It can also be a named logical vector
to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{nrow}{integer - the number of rows for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{nrow}).}

\item{ncol}{integer - the number of columns for the correlation labels if
grouping aesthetics are present (defaults to \code{NULL} - automatic setting of
\code{ncol}).}

\item{digits}{integer: Number of digits the correlations are rounded to
(defaults to 2).}

\item{corr_method}{character string with the correlation method passed to
\link[=stats]{cor()}. Can be one of \code{"pearson"}, \code{"kendall"} and \code{"spearman"}.
Defaults to \code{"pearson"} (or is inherited from the setting in \code{\link[=ggcorrm]{ggcorrm()}}).}

\item{squeeze}{numeric between 0 an 1. Proportion of the facet width/height
the facet labels are restricted to when multiple labels are present
(defaults to 0.7 - labels extend over 70\% of the extent of the plot).}

\item{byrow}{logical. Should the correlation labels in plots with multiple
groups be filled by rows (\code{byrow = TRUE}) or by columns (\code{byrow = FALSE})?
Note that the actual number of rows or columns that are filled with values
can be below the specified value of \code{nrow} or \code{ncol} when less rows/columns
than specified are needed to reach the total number of groups. Defaults to
\code{TRUE}.}

\item{...}{additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_corrtext()} is used to compute bivariate correlations and
appropriate positions of text labels indicating correlation strength for
the facets of \link{ggcorrm} plots.
}
\details{
\code{stat_corrtext()} computes the correlation between variables in the
facets of \code{ggcorrm} plots and places text labels indicating the strength of
correlation in appropriate positions within the facets.
}
\seealso{
\code{\link[=lotri_corrtext]{lotri_corrtext()}} and \code{\link[=utri_corrtext]{utri_corrtext()}}
}
