% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_bin.R
\name{stat_dia_bin}
\alias{stat_dia_bin}
\title{FUNCTION_TITLE}
\usage{
stat_dia_bin(
  mapping = NULL,
  data = NULL,
  geom = "rect",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  lower = 0.25,
  upper = 1,
  barwidth = 0.9,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'rect'}

\item{position}{PARAM_DESCRIPTION, Default: 'identity'}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}

\item{lower}{PARAM_DESCRIPTION, Default: 0.25}

\item{upper}{PARAM_DESCRIPTION, Default: 1}

\item{barwidth}{PARAM_DESCRIPTION, Default: 0.9}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
