% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_names.R
\name{stat_dia_names}
\alias{stat_dia_names}
\title{FUNCTION_TITLE}
\usage{
stat_dia_names(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  y_pos = 0.2,
  ...
)
}
\arguments{
\item{mapping}{PARAM_DESCRIPTION, Default: NULL}

\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{geom}{PARAM_DESCRIPTION, Default: 'text'}

\item{position}{PARAM_DESCRIPTION, Default: 'identity'}

\item{show.legend}{PARAM_DESCRIPTION, Default: NA}

\item{inherit.aes}{PARAM_DESCRIPTION, Default: TRUE}

\item{y_pos}{PARAM_DESCRIPTION, Default: 0.2}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
