% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dia_names.R
\name{stat_dia_names}
\alias{stat_dia_names}
\title{Positions for variable names in ggcorrm plots}
\usage{
stat_dia_names(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  y_pos = 0.2,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{y_pos}{numeric between 0 and 1 specifying the relative position of the
text labels along the x axis (defaults to 0.2).}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{layer()}}}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_dia_names()} is used by
     \code{\link[dia_names]{dia_names()}} to compute the positions of layer
     names in the diagonal panels of \code{\link{ggcorrm}} plots.
}
\details{
\code{stat_dia_names()} uses the range of variables on the diagonal
    facets of \code{\link{ggcorrm}} plots to compute appropriate positions
    for text labels of variable names.
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer}},
  \code{\link{dia_names}}
}
