% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heatcircle.R
\name{stat_heatcircle}
\alias{stat_heatcircle}
\title{Compute circles whose size represents correlation strength}
\usage{
stat_heatcircle(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  corr_method = "pearson",
  rmin = 0.1,
  rmax = 0.9,
  scale_by = c("area", "radius"),
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by
\code{\link[ggplot2:aes]{aes()}}. \code{x} and \code{y} are set
automatically and must not be changed,  but all other aesthetics
may be manipulated. Defaults to \code{NULL} (use standard mapping).}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{corr_method}{character string with the correlation method passed
to \code{\link[stats]{cor}}. Can be one of "pearson", "kendall" and
"spearman". Defaults to "pearson" (or is inherited from the setting
in \code{\link[ggcorrm]{ggcorrm()}}).}

\item{rmin}{numeric (0 - 1). Minimum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.1.}

\item{rmax}{numeric (0 - 1). Maximum radius for the "heat circles" (as a
proportion of panel width). Defaults to 0.9.}

\item{scale_by}{character.  \code{scale_by = "area"} scales the area of the
circles by correlation strength, \code{scale_by = "radius"} scales by
radius. Defaults to "area".}

\item{...}{additional arguments passed to
\code{\link[ggplot2:layer]{ggplot2::layer}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_heatcircle()} is used to compute bivariate correlations
    and display them in \code{\link{ggcorrm}} plots in form of circles whose color
    and size indicate correlation strength.
}
\details{
\code{stat_heatcircle()} computes the outlines of circles that
   are centered around the midpoint of \code{gcorrm} facets and whose radius
   or area is modified in dependence of correlation strength. \code{rmin} and
   \code{rmax} allow to modify the range in which the radii of the circles may
   vary. The circles are displayed using \code{\link[ggplot2]{geom_ribbon}}.
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer}},
  \code{\link{geom_heatcircle}}
}
