% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_heatmap.R
\name{stat_heatmap}
\alias{stat_heatmap}
\title{Compute correlation heatmaps for ggcorrm plots}
\usage{
stat_heatmap(
  mapping = NULL,
  data = NULL,
  geom = "rect",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  corr_method = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \link[ggplot2:aes]{aes} that
are passed on to subsequent layers. \code{x} and \code{y} are set automatically and
must not be changed,  but all other aesthetics may be manipulated. Defaults
to \code{NULL} (use standard \code{ggcorrm} mapping).}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
FALSE (the default) never includes, TRUE always includes, and NA includes
only if aesthetics are mapped. It can also be a named logical vector
to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{corr_method}{character string with the correlation method passed to
\link[=stats]{cor()}. Can be one of \code{"pearson"}, \code{"kendall"} and \code{"spearman"}.
Defaults to \code{"pearson"} (or is inherited from the setting in \code{\link[=ggcorrm]{ggcorrm()}}).}

\item{...}{additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\value{
An object of class \code{Layer}.
}
\description{
\code{stat_heatmap()} is used to compute bivariate correlations and
display them as correlation heat maps / point labels in the facets of
\link{ggcorrm} plots.
}
\details{
\code{stat_heatmap()} computes the correlation between variables in the
facets of \code{ggcorrm} plots and returns the aesthetics necessary to either
color the entire facet by correlation strength (\code{geom = "rect"}) or place a
single point in the midpoint of each facet (\code{geom = "point"}).
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}},
\code{\link[=add_heatmap]{add_heatmap()}}
}
